Unblock docker REST API registry PUT requests
According to the Apache config documentation, the <IfDefine> container is evaluated at server startup and restart. If the condition is true, then the enclosed directives will be applied to all requests. Since FRONTEND_BROWSE_ONLY_MODE is set by default, only GET requests were passed to the Docker Registry but no PUT requests. Therefore the command "docker push <image>" throws an 403 error when using this docker registry frontend as a proxy for the docker registry.